+++ src/features/logger.ts
@@ -1,32 +1,50 @@
-import { IpcMain, IpcRenderer } from "electron";
+import { IpcMain, ipcRenderer, ipcMain } from "electron";
 import { globalEvents } from "../constants/globalEvents";
 
 export class Logger {
-  /**
-   *
-   * @param ipcRenderer renderer IPC client so we can send messages to the main thread
-   */
-  constructor(private ipcRenderer: IpcRenderer) {}
-
   /**
    * Subscribe to watch for logs from the IPC client
    * @param ipcMain main thread IPC client so we can subscribe to events
    */
   public static watch(ipcMain: IpcMain) {
-    ipcMain.on(globalEvents.log, (event, content, object) => {
-      console.log(content, JSON.stringify(object, null, 2));
+    ipcMain.on(globalEvents.log, (event, message) => {
+      const { content, object } = message;
+      this.logToSTDOut(content, object);
     });
   }
-
   /**
    * Log content to STDOut
    * @param content
    * @param object js(on) object that will be prettyPrinted
    */
-  public log(content: string, object: object = {}) {
-    if (this.ipcRenderer) {
-      this.ipcRenderer.send(globalEvents.log, { content, object });
+  public static log(content: string, object: object = {}) {
+    if (ipcRenderer) {
+      ipcRenderer.send(globalEvents.log, { content, object });
+    } else {
+      ipcMain.emit(globalEvents.log, { content, object });
     }
-    console.log(`${content} \n ${JSON.stringify(object, null, 2)}`);
+    this.logToSTDOut(content, object);
+  }
+
+  /**
+   * Log content to STDOut and use the provided alert function to alert
+   * @param content
+   * @param object js(on) object that will be prettyPrinted
+   */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  public static alert(content: string, object: any = {}, alert?: (msg: string) => void) {
+    Logger.log(content, object);
+    if (alert) {
+      alert(`${content} \n\nwith details: \n${JSON.stringify(object, null, 2)}`);
+    }
+  }
+
+  /**
+   * Log to STDOut
+   * @param content
+   * @param object
+   */
+  private static logToSTDOut(content: string, object = {}) {
+    console.log(content, Object.keys(object).length > 0 ? JSON.stringify(object, null, 2) : "");
   }
 }