+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [next]
+
+- [DEV]:
+  - Logger is now static and will automatically call either ipcRenderer or ipcMain
+
 ## 5.5.0
 
 - ListenBrainz integration added (thanks @Mar0xy)